fix: use BigDecimal for similarity scores per project rules
- ScoredChunk.similarityScore: double → BigDecimal
- Citation.score: double → BigDecimal
- PgVectorStoreService.toScoredChunk: wrap doc.getScore() with BigDecimal.valueOf()
- PromptContextBuilderService: Comparator.comparingDouble → Comparator.comparing
- Update log format strings from %.3f to %s for BigDecimal scores
- Fix all test helpers and assertions to use BigDecimal.valueOf()

+++ src/main/java/com/test/rag/model/Citation.java
@@ -1,9 +1,11 @@
 package com.test.rag.model;
 
+import java.math.BigDecimal;
+
 public record Citation(
         int ref,
         String filename,
         int chunkIndex,
-        double score,
+        BigDecimal score,
         String chunkText
 ) {}
